+++ CHANGELOG.md
@@ -10,10 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added method `mat_id` to `OutputAssignments`.
 * Added method `update_weights` to `skinning.Weights`.
 * Added method `add_influences` to `skinning.SkinWeights`.
+* Added field `morph_blend_target` to `vertex.VertexBuffer` for base morph attributes.
+* Added base morph attributes `Position2`, `Normal4`, `OldPosition`, and `Tangent2` to `vertex.AttributeType`.
 
 ### Changed
 * Adjusted the signature of `animation.Track.sample_` methods to also take the frame count.
 * Adjusted fields to use references to make nested property access like `root.buffers.vertex_buffers` work as expected.
+* Changed `MorphTarget.tangent_deltas` to `MorphTarget.tangents` to better reflect in game data.
+* Changed `MorphTarget.normal_deltas` to `MorphTarget.normals` to better reflect in game data.
 
 ### 0.5.1 - 2024-04-29
 ### Fixed